+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/text/package-info.java
@@ -21,6 +21,7 @@
  * <ul>
  * <li>{@link is.codion.swing.common.ui.component.text.NumberField#CONVERT_GROUPING_TO_DECIMAL_SEPARATOR}
  * <li>{@link is.codion.swing.common.ui.component.text.TextFieldBuilder#SELECT_ALL_ON_FOCUS_GAINED}
+ * <li>{@link is.codion.swing.common.ui.component.text.TemporalField#ADJUSTABLE}
  * </ul>
  */
 @org.jspecify.annotations.NullMarked